requirements.txt
spotipy==2.23.0
requests==2.31.0
python-dotenv==1.0.1
ytmusicapi==1.3.0